now better decompress zip files, better handle skip_columns

+++ convert_dicoms_to_niftis.py
@@ -10,6 +10,9 @@ from anima_utils import *
 parser = shanoir_downloader.create_arg_parser("Convert DICOMs to Niftis")
 
 parser.add_argument('-d', '--dicoms', required=True, help='Path to the folder containing the dicoms.')
+parser.add_argument('-of', '--output_folder', required=False, help='Path to the output folder (only used when using from_folders).')
+parser.add_argument('-ff', '--from_folders', default=False, action='store_true', help='Convert from folders instead of zip archives.')
+parser.add_argument('-kd', '--keep_dicom', default=False, action='store_true', help='Keep dicoms after conversion.')
 
 args = parser.parse_args()
 
@@ -39,20 +42,22 @@ def image_is_readable(image_path):
     except:
         return False
 
-def convert_dicom_to_nifti(dicom_directory, conversion_info):
-    dicom_parent = dicom_directory.parent
-    if str(dicom_parent) in [ci['path'] for ci in conversion_info]:
-        return dicom_parent
-    
+def convert_dicom_to_nifti(dicom_directory, output_folder):
+    output_folder.mkdir(parents=True, exist_ok=True)
+
     image_converted = False
     
-    dicom2nifti_command = lambda: dicom2nifti.convert_directory(dicom_directory, dicom_parent, compression=True, reorient=True)
-    mricron_dcm2niix_command = lambda: call(['/home/amasson/apps/dcm2nii/mricron/dcm2niix', '-z', 'y', '-w', '1', '-o', dicom_parent, dicom_directory])
-    dcm2niix_command = lambda: call(['dcm2niix', '-z', 'y', '-o', dicom_parent, dicom_directory])
-    mcverter_command = lambda: call(['mcverter', '-o', dicom_parent, '-f', 'nifti', '-n', dicom_directory])
+    dicom2nifti_command = lambda: dicom2nifti.convert_directory(dicom_directory, output_folder, compression=True, reorient=True)
+    mricron_dcm2niix_command = lambda: call(['/home/amasson/apps/dcm2nii/mricron/dcm2niix', '-z', 'y', '-w', '1', '-o', output_folder, dicom_directory])
+    dcm2niix_command = lambda: call(['dcm2niix', '-z', 'y', '-o', output_folder, dicom_directory])
+    mcverter_command = lambda: call(['mcverter', '-o', output_folder, '-f', 'nifti', '-n', dicom_directory])
 
     commands = [dicom2nifti_command, mricron_dcm2niix_command, dcm2niix_command, mcverter_command]
     commandNames = ['dicom2nifti', 'mricronDcm2niix', 'dcm2niix', 'mcverter']
+
+    conversion_tool = None
+    reconverted = None
+    image_converted = None
 
     for i, command in enumerate(commands):
         conversion_tool = commandNames[i]
@@ -61,7 +66,7 @@ def convert_dicom_to_nifti(dicom_directory, conversion_info):
             command()
         except:
             pass
-        image_path = get_first_nifti(dicom_parent)
+        image_path = get_first_nifti(output_folder)
         if image_path is None: continue
         image_converted = image_is_readable(image_path)
         if image_converted: break
@@ -70,22 +75,37 @@ def convert_dicom_to_nifti(dicom_directory, conversion_info):
         image_converted = image_is_readable(image_path)
         if image_converted: break
 
-    if image_converted:
-        shutil.rmtree(str(dicom_directory))
-    else:
-        print('ERROR: dicom could not be converted!', dicom_directory)
+    return image_converted, conversion_tool, reconverted, image_converted
+
+def convert_dicom_to_nifti_if_needed(dicom_directory, output_folder, keep_dicom, conversion_info):
+
+    if str(output_folder) in [ci['path'] for ci in conversion_info]:
+        return output_folder
     
-    conversion_info.append({'path': str(dicom_parent), 'conversion_tool': conversion_tool, 'reconverted': reconverted, 'converted': image_converted})
+    image_converted, conversion_tool, reconverted, image_converted = convert_dicom_to_nifti(dicom_directory, output_folder)
+
+    if not keep_dicom:
+        if image_converted:
+            shutil.rmtree(str(dicom_directory))
+        else:
+            print('ERROR: dicom could not be converted!', dicom_directory)
+
+    conversion_info.append({'path': str(output_folder), 'conversion_tool': conversion_tool, 'reconverted': reconverted, 'converted': image_converted})
     save_conversion_tools(conversion_info)
-    return dicom_parent
+
+    return output_folder
 
 for dicom in sorted(list(dicoms.iterdir())):
-    # Extract the zip file
-    zip_files = list(dicom.glob('*.zip'))
-    for dicom_zip in zip_files:
-        print(dicom_zip)
-        logging.info(f'    Extracting {dicom_zip}...')
-        dicom_folder = dicom_zip.parent / dicom_zip.stem
-        dicom_folder.mkdir(exist_ok=True)
-        shutil.unpack_archive(str(dicom_zip), str(dicom_folder))
-        convert_dicom_to_nifti(dicom_folder, conversion_info)
+    if args.from_folders:
+        print('Converting', dicom)
+        convert_dicom_to_nifti_if_needed(dicom, Path(args.output_folder) / f'{dicom.name}', args.keep_dicom, conversion_info)
+    else:
+        # Extract the zip file
+        zip_files = list(dicom.glob('*.zip'))
+        for dicom_zip in zip_files:
+            print(dicom_zip)
+            logging.info(f'    Extracting {dicom_zip}...')
+            dicom_folder = dicom_zip.parent / dicom_zip.stem
+            dicom_folder.mkdir(exist_ok=True)
+            shutil.unpack_archive(str(dicom_zip), str(dicom_folder))
+            convert_dicom_to_nifti_if_needed(dicom_folder, dicom_folder.parent, args.keep_dicom, conversion_info)
